MBS-11503 Move TestCase.kt to the test-module
