+++ subprojects/test-runner/client/src/main/kotlin/com/avito/runner/scheduler/listener/ReportProcessor.kt
@@ -2,7 +2,7 @@ package com.avito.runner.scheduler.listener
 
 import com.avito.report.model.AndroidTest
 import com.avito.runner.scheduler.logcat.LogcatAccessor
-import com.avito.runner.service.model.TestCase
+import com.avito.test.model.TestCase
 
 /**
  * todo remove in favour of TestArtifactsProcessorImpl after uploadAllArtifacts flag removed
