+++ subprojects/test-runner/client/src/main/kotlin/com/avito/runner/scheduler/runner/TestRunner.kt
@@ -2,7 +2,7 @@ package com.avito.runner.scheduler.runner
 
 import com.avito.android.Result
 import com.avito.runner.scheduler.runner.model.TestRunnerResult
-import com.avito.runner.service.model.TestCase
+import com.avito.test.model.TestCase
 
 internal interface TestRunner {
