+++ subprojects/test-runner/client/src/main/kotlin/com/avito/runner/scheduler/runner/TestRunnerImpl.kt
@@ -15,9 +15,9 @@ import com.avito.runner.scheduler.runner.model.TestRunResult
 import com.avito.runner.scheduler.runner.model.TestRunnerResult
 import com.avito.runner.scheduler.runner.scheduler.TestExecutionScheduler
 import com.avito.runner.service.DeviceWorkerPool
-import com.avito.runner.service.model.TestCase
 import com.avito.test.model.DeviceName
 import com.avito.time.millisecondsToHumanReadableTime
+import com.avito.test.model.TestCase
 import kotlinx.coroutines.coroutineScope
 
 internal class TestRunnerImpl(
