+++ subprojects/test-runner/client/src/main/kotlin/com/avito/runner/scheduler/runner/model/TestRunnerResult.kt
@@ -1,7 +1,7 @@
 package com.avito.runner.scheduler.runner.model
 
 import com.avito.runner.service.model.DeviceTestCaseRun
-import com.avito.runner.service.model.TestCase
+import com.avito.test.model.TestCase
 
 internal data class TestRunnerResult(
     val runs: Map<TestCase, List<DeviceTestCaseRun>>
