+++ subprojects/test-runner/client/src/test/kotlin/com/avito/runner/scheduler/metrics/model/TestKey.kt
@@ -1,7 +1,7 @@
 package com.avito.runner.scheduler.metrics.model
 
-import com.avito.runner.service.model.TestCase
 import com.avito.test.model.DeviceName
+import com.avito.test.model.TestCase
 import com.avito.test.model.TestName
 
 // todo move to :service
