+++ subprojects/test-runner/client/src/testFixtures/kotlin/com/avito/runner/scheduler/runner/model/StubTestRunRequest.kt
@@ -1,9 +1,9 @@
 package com.avito.runner.scheduler.runner.model
 
-import com.avito.runner.service.model.TestCase
 import com.avito.runner.service.model.createStubInstance
 import com.avito.runner.service.worker.device.model.DeviceConfiguration
 import com.avito.runner.service.worker.device.model.createStubInstance
+import com.avito.test.model.TestCase
 
 internal fun TestRunRequest.Companion.createStubInstance(
     testCase: TestCase = TestCase.createStubInstance(),
