+++ subprojects/test-runner/service/build.gradle.kts
@@ -15,6 +15,7 @@ dependencies {
     api(project(":common:time"))
     api(project(":test-runner:command-line-executor"))
     api(project(":gradle:process"))
+    api(project(":test-runner:test-model"))
 
     implementation(project(":common:result"))
     implementation(project(":common:problem"))
