+++ subprojects/test-runner/service/src/main/kotlin/com/avito/runner/service/listener/TestListener.kt
@@ -1,9 +1,9 @@
 package com.avito.runner.service.listener
 
 import com.avito.android.Result
-import com.avito.runner.service.model.TestCase
 import com.avito.runner.service.model.TestCaseRun
 import com.avito.runner.service.worker.device.Device
+import com.avito.test.model.TestCase
 import java.io.File
 
 interface TestListener {
