+++ subprojects/test-runner/service/src/main/kotlin/com/avito/runner/service/model/intention/InstrumentationTestRunAction.kt
@@ -1,6 +1,6 @@
 package com.avito.runner.service.model.intention
 
-import com.avito.runner.service.model.TestCase
+import com.avito.test.model.TestCase
 
 data class InstrumentationTestRunAction(
     val test: TestCase,
