+++ subprojects/test-runner/service/src/testFixtures/kotlin/com/avito/runner/service/model/intention/StubInstrumentationTestRunAction.kt
@@ -1,7 +1,7 @@
 package com.avito.runner.service.model.intention
 
-import com.avito.runner.service.model.TestCase
 import com.avito.runner.service.model.createStubInstance
+import com.avito.test.model.TestCase
 
 fun InstrumentationTestRunAction.Companion.createStubInstance(
     testCase: TestCase = TestCase.createStubInstance(),
