+++ subprojects/test-runner/test-model/src/main/kotlin/com/avito/test/model/TestCase.kt
@@ -1,0 +1,15 @@
+package com.avito.test.model
+
+import java.io.Serializable
+
+public data class TestCase(
+  val name: TestName,
+  val deviceName: DeviceName
+) : Serializable {
+
+    override fun toString(): String {
+        return "$name.$deviceName"
+    }
+
+    public companion object
+}
